+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity"
+        <activity android:name=".ui.main.MainActivity"
             android:screenOrientation="portrait">
         </activity>
