--- app/src/main/java/com/tkachuk/cardholderapp/MainActivity.kt
@@ -1,31 +1,0 @@
-package com.tkachuk.cardholderapp
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-
-class MainActivity : AppCompatActivity() {
-
-    private val mainPresenter: MainPresenter  = MainPresenter()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        val inflater = menuInflater
-        inflater.inflate(R.menu.menu, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
-        return if (item != null && item.itemId == R.id.logout) {
-            mainPresenter.logout()
-            finish()
-            true
-        }
-        else super.onOptionsItemSelected(item)
-    }
-}
